adding custom response for validation

+++ HRLeaveManagement.Application/Features/LeaveAllocations/Handlers/Commands/LeaveAllocationDeleteCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using HRLeaveManagement.Application.Exceptions;
 using HRLeaveManagement.Application.Features.LeaveAllocations.Requests.Commands;
 using HRLeaveManagement.Application.Persistence.Contracts;
+using HRLeaveManagement.Domain;
 using MediatR;
 
 namespace HRLeaveManagement.Application.Features.LeaveAllocations.Handlers.Commands
@@ -20,6 +22,7 @@ namespace HRLeaveManagement.Application.Features.LeaveAllocations.Handlers.Comma
         public async Task Handle(LeaveAllocationDeleteCommand request, CancellationToken cancellationToken)
         {
             var leaveAllocation = await _leaveAllocationRepository.GetById(request.Id);
+            if (leaveAllocation == null) { throw new NotFoundException(nameof(LeaveAllocation), request.Id); };
             await _leaveAllocationRepository.Delete(leaveAllocation);
 
         }
