+++ HRLeaveManagement.Application/Features/LeaveRequests/Handlers/Commands/LeaveRequestDeleteCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using HRLeaveManagement.Application.Exceptions;
 using HRLeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
 using HRLeaveManagement.Application.Persistence.Contracts;
+using HRLeaveManagement.Domain;
 using MediatR;
 
 namespace HRLeaveManagement.Application.Features.LeaveRequests.Handlers.Commands
@@ -18,6 +20,7 @@ namespace HRLeaveManagement.Application.Features.LeaveRequests.Handlers.Commands
         public async Task Handle(LeaveRequestDeleteCommand request, CancellationToken cancellationToken)
         {
             var leaveRequset = await _leaveRequestRepository.GetById(request.Id);
+            if (leaveRequset == null) { throw new NotFoundException(nameof(LeaveRequest), request.Id); };
             await _leaveRequestRepository.Delete(leaveRequset);
 
         }
