+++ HRLeaveManagement.Application/Features/LeaveTypes/Handlers/Commands/LeaveTypeDeleteCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using HRLeaveManagement.Application.Exceptions;
 using HRLeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
 using HRLeaveManagement.Application.Persistence.Contracts;
+using HRLeaveManagement.Domain;
 using MediatR;
 
 namespace HRLeaveManagement.Application.Features.LeaveTypes.Handlers.Commands
@@ -18,6 +20,7 @@ namespace HRLeaveManagement.Application.Features.LeaveTypes.Handlers.Commands
         public async Task Handle(LeaveRequestDeleteCommand request, CancellationToken cancellationToken)
         {
             var leaveType = await _leaveTypeRepository.GetById(request.Id);
+            if (leaveType == null) { throw new NotFoundException(nameof(LeaveType), request.Id); };
             await _leaveTypeRepository.Delete(leaveType);
 
         }
